% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ica}
\alias{ica}
\title{ICA Probabilistic Cortical Parcellation}
\usage{
ica()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\description{
Brain atlas for the probabilistic independent component analysis (ICA)
cortical parcellation with 163 regions. Contains 2D polygon geometry
for \code{\link[ggseg:ggbrain]{ggseg::geom_brain()}}.
}
\examples{
ica()
}
\references{
Beckmann CF, Smith SM (2004). Probabilistic Independent
Component Analysis for Functional Magnetic Resonance Imaging. \emph{IEEE
Transactions on Medical Imaging}, 23(2):137-152.
\doi{10.1109/TMI.2003.822821}
}
\concept{ggseg_atlases}
